Ensure `Telemetry` is not exposed
