+++ js/src/tests/vercel/wrapper/methods.vitesttest.ts
@@ -1824,6 +1824,7 @@ describe("wrapAISDK", () => {
       await wrappedMethods.generateText({
         model: mockLangModel,
         prompt: "What's the weather?",
+        // @ts-expect-error - experimental_output is not typed
         experimental_output: ai.Output.object({
           schema: z.object({
             city: z.string(),
@@ -1906,6 +1907,7 @@ describe("wrapAISDK", () => {
       const result = wrappedMethods.streamText({
         model: mockLangModel,
         prompt: "What's the weather?",
+        // @ts-expect-error - experimental_output is not typed
         experimental_output: ai.Output.object({
           schema: z.object({
             city: z.string(),
